Create Event Sync Service (#114)
* Create Event Sync Service

Create Event Sync as a service

Checks of the event is already present (via hash) or if it needs to
be updated.

Imports images from events site for persistence.

Create rake task to trigger service

* Refactors tests for FuzzyFind and for Event sync

Turns out that when you stub a call with a test double, it stays
stubbed. This borke integration tests in otherparts of the suite.

Use allow rather than double to localize the mocking.