All forms successfully emailed are persisted (#620)
As form field vary greatly between form, we cannot persist them with a
column for each field. Instead we serialize as JSON the attributes into
`form_attributes`.

Because forms may contain sensitve data, we encrypt the form_attributes
before they are saved to the DB, and are decrypted on access via the
model, assuming  LOCKBOX_MASTER_KEY nev var value is the same it was at
the time of encryption.

This PR also adds some tests to a few forms to ensure the form renders,
can send email (in a test environment), and persiststhe submitted data
to the db. There are still several forms that require tests.