Refactor Finding Aids query logic
Move the majority of filtering logic from controller to model.

Add scopes to the Finding Aid model to handle filtering by collection or
subject.

Current implementation queries every fiding aid individually to get
details used for building filters, which is a major performance hit. This decreases the number of queries we make to the DB for each index page dramatically, and this improves performance.