Add verbose logging, force option for event sync
Adds much more granular logging of event sync jobs.

Adds a force option, which causes all events in the feed to overwrite
existing events, not skipping when a matching hash is found. Useful when
code changes are made to the events sync that would not be picked up
because of the content_hash match.

The sync button in the Events dashboard now uses the force sync, so this
can be used manually after deploys the alter this code.