[BL-1017] Fix dummy_range field bug.
Blacklight Raange Limit adds search_field=dummy_range if dummy_range is
not set in the query.  This seems to be an outdated requirement that I'm
trying to get removed and thus fix issue, but in the mean time we can
fix our issue simply by making sure we set the param searc_field=all
when submitting queries to the catalog.

REF: https://github.com/projectblacklight/blacklight_range_limit/pull/109
REF: BL-1017