MAN-1484 scrc forms investigation (#3995)
Creates new AV request form that does not send email, but writes to db
and makes downloadable CSV file available. This form should not be
listed on the forms page and only linked to from the SCRC homepage.

This includes some work to fix failing tests for buildings drafts,
unrelated, unexpected, but needed.